### 2025-09-26 #9083 @editorialbot
[PRE REVIEW]: EAM QuickView: A simulation data visualizer for the E3SM Atmosphere Model
**Submitting author:** @ayenpure (<a href="http://orcid.org/0000-0002-7403-7661">Abhishek Yenpure</a>)
**Repository:** https://github.com/ayenpure/QuickView
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.0
