### 2025-10-01 #9083 @kthyng
@ayenpure
* can this be released on PyPI or conda-forge?
* are there tests?
