### 2025-11-14 #9083 @ayenpure
@kthyng I'm extremely sorry for my slow response -- for most of October I was on parental leave and I'm still getting caught up on things. Right before I went on my leave we've had a couple big changes to the project.
1. We migrated the project owner from me to my org -- Kitware so there is a repository change
2. We made changes to how the app is deployed -- it can now be installed using `pip`.

I am wondering what would be the best course of action for us to update the repository.
Should we resubmit? of is there a good way to update that information with this current review request?
