### 2025-11-14 #9083 @ayenpure
@kthyng After discussion with my colleagues, here is an updated response.

We want to keep the current repository for the atmosphere component for now, the new repository is still in the works to generalize it for any earth system tool.

In that context the answers to your questions:

1. The app can locally be installed using `pip` -- this needs the user/developer to clone the repo and `pip install .` or `pip install -e .`  Our intention still is for end users to use the packages binaries as an app, and only for the developers to use the pythonic way.

2. Answer remains the same
