### 2025-11-14 #9083 @kthyng
@ayenpure Ah, I hope your parental leave went well! 👶

1. Since one of the points of publishing with JOSS is to ensure good practices for new users to a package including development, you should make the package available for users both as an app and as a developer. Any reviewers will need to install it.
2. It looks like your CI workflow tests just make sure the package installs and sets up, is that correct?
