### 2025-11-14 #9083 @ayenpure
> [@ayenpure](https://github.com/ayenpure) Ah, I hope your parental leave went well! 👶
>
> 1. Since one of the points of publishing with JOSS is to ensure good practices for new users of a package, including development, you should make the package available to users both as an app and as a developer. Any reviewers will need to install it.

The reviewers could still install the app and use it. We recommend installing the compiled binaries, as we package the application as a desktop app. And yes, technically everyone can use the python application. They can still `pip install .` and install locally. The only part we're missing is that we don't publish it on `conda-forge` or `pip` as we depend on ParaView, which has its own Python subsystem. It's challenging to ensure agreement between the Python environment used to publish and the user's Python environment when published to pip or conda-forge.

> 2. It looks like your CI workflow tests that the package installs and sets up. Is that correct?

That is correct!
