### 2026-01-27 6610970
Restructure JOSS paper to comply with updated submission guidelines
  - Add required "State of the field"
  section comparing to ncview/ncvis
  - Add "Research impact statement"
  section for E3SM community adoption
  - Add "AI usage disclosure" section
  (no AI was used)
  - Rename "Technical Details" to
  "Software design"
  - Remove non-standard "Overall
  strategy" section, redistributing
  content
