### 2026-01-28 #9083 @trontrytel
> [@trontrytel](https://github.com/trontrytel) , [@mnlevy1981](https://github.com/mnlevy1981), [@KMFleischer](https://github.com/KMFleischer) would any of you be interested in reviewing this paper and software package for the journal of open source software

Apologies for the delay - I was on travel. I can be a reviewer if you still need it!
