### 2026-01-28 #9083 @KMFleischer
I‘m very sorry but I do not have enough time to work on this. Regards,KMFleischerAm 28.01.2026 um 22:24 schrieb Anna Jaruga ***@***.***>:﻿trontrytel left a comment (openjournals/joss-reviews#9083)

@trontrytel , @mnlevy1981, @KMFleischer would any of you be interested in reviewing this paper and software package for the journal of open source software

Apologies for the delay - I was on travel. I can be a reviewer if you still need it!
