### 2026-01-29 19b928b
revise JOSS manuscript
- moved and revised some paragraphs;
- disclosed use of ChatGPT for manuscirpt editing.
