+++ docs/JOSS/paper.md
@@ -56,60 +56,34 @@ in visual analytics.
 EAM, like many other simulation codes used in Earth system sciences, solves a
 complex set of equations and writes out results for a large number of physical
 quantities in the form of NetCDF files. After a simulation is completed and
-before more focused and detailed analyses are performed, it is often useful to
+before more detailed analyses are performed, it is often useful to
 obtain a first impression of the characteristic values of the simulated
 quantities and their geographical distributions. The atmosphere modelers' wish
 for tools that can simultaneously present multiple variables has become more
-prominent in the past years due to the rapid increase of model complexity in
+prominent in recent years  due to the rapid increase of model complexity in
 terms of the number of equations solved by the numerical models and the number
 of variables typically archived in the NetCDF files. For example, to inspect the
-simulated aerosol life cycles, it may be useful to examine the concentrations of
+simulated aerosol life cycles, it is very useful to examine the concentrations of
 multiple chemical components in different particle size ranges (see Figure 1).
 
 ![Figure 1: Aerosol concentrations simulated by EAMv2.](images/fig_EAMv2_aerosols.png)
 
-Since the current work is focused on facilitating the development of EAM, our
-app is designed to work with the cubed sphere horizontal mesh and the
-pressure-based terrain following vertical coordinate. The adaptation of EAM
-QuickView for other models and meshes can be done by updating the ParaView
-Reader in the app.
-
 # State of the field
 
-Several tools exist for quick visualization of NetCDF files from Earth system
-models. The tool [ncview](https://cirrus.ucsd.edu/ncview/) developed by
-[@DavidWPierce:1993] has been used widely in Earth system modeling communities
-for quick and easy, push-button inspections of NetCDF files. The newer tool
-[ncvis](https://github.com/SEATStandards/ncvis) developed by [@PaulUllrich:2022]
-is inspired by ncview but can also handle data on unstructured meshes, making it
-usable out-of-the-box for many different models and mesh types.
+Several tools exist for quick visualization of NetCDF files produced by Earth system
+models. [ncview](https://cirrus.ucsd.edu/ncview/) [@DavidWPierce:1993]
+has been widely used in Earth system modeling communities
+for rapid, push-button inspections of NetCDF files. The newer tool
+[ncvis](https://github.com/SEATStandards/ncvis) developed by [@PaulUllrich:2022] and
+inspired by ncview extends this capability to data defined on unstructured meshes, making it usable out of the box for a broad range of models and mesh types.
+While EAM QuickView is also designed for rapid inspections, it was developed to address the need to simultaneously display multiple variables in a single viewport and to support flexible rearrangement of variables during scientific reasoning. To facilitate this more sophisticated analysis workflow, QuickView also includes session persistence via state files that can be saved and reloaded, allowing users to resume analysis and achieve provenance.
 
-EAM QuickView was built rather than contributing to existing projects for
-several reasons. First, ncview and ncvis are designed to display one variable at
-a time, whereas EAM QuickView can simultaneously present multiple variables in a
-single viewport with drag-and-drop arrangement. Second, EAM QuickView provides
-session persistence through state files that can be saved and reloaded,
-preserving the visualization configuration for later use. Third, by building on
-ParaView's visualization engine, EAM QuickView gains access to ParaView's large
-collection of analysis capabilities and built-in ability to handle big data
-through parallel processing, which is imperative for global atmospheric models
-in the era of high-resolution modeling.
[…]
