### 2026-01-29 10f7bca
minor edit in docs/JOSS/paper.md
