+++ docs/JOSS/paper.md
@@ -78,7 +78,7 @@ for rapid, push-button inspections of NetCDF files. The newer tool
 inspired by ncview extends this capability to data defined on unstructured meshes, making it usable out of the box for a broad range of models and mesh types.
 While EAM QuickView is also designed for rapid inspections, it was developed to address the need to simultaneously display multiple variables in a single viewport and to support flexible rearrangement of variables during scientific reasoning. To facilitate this more sophisticated analysis workflow, QuickView also includes session persistence via state files that can be saved and reloaded, allowing users to resume analysis and achieve provenance.
 
-To leverage state-of-the-art capabilities in visual analytics, EAM QuickView uses [ParaView](https://www.paraview.org/) [@Ahrens:2005; @Henderson:2007] as its visualization engine, thereby inheriting ParaView’s extensive collection of analysis algorithms and its built-in support for parallel processing of large data sets—capabilities that are indispensable for atmospheric models in the era of high-resolution modeling. By design, EAM QuickView departs from ParaView’s general-purpose UI which presents a steep learning curve and has repeatedly proven to be a barrier to effective use for scientists unfamiliar with the visual analytics paradigm. Instead, EAM QuickView provides a UI explicitly tailored to the scientific language and reasoning workflows of atmospheric modeling. The UI is intentionally focused on a single, well-defined task, namely rapid inspection of simulation output. In this sense, EAM QuickView represents the first prototype in a family of ParaView-based analysis UIs that is purpose-built for atmospheric applications. Additional members of this tool family currently under development target workflows such as rapid comparison of multiple simulations and flexible slicing of three-dimensional fields using arbitrary planes and bounding boxes.
+To leverage state-of-the-art capabilities in visual analytics, EAM QuickView uses [ParaView](https://www.paraview.org/) [@Ahrens:2005; @Henderson:2007] as its visualization engine, thereby inheriting ParaView’s extensive collection of analysis algorithms and its built-in support for parallel processing of large data sets—capabilities that are indispensable for atmospheric models in the era of high-resolution modeling. By design, EAM QuickView departs from ParaView’s general-purpose UI which presents a steep learning curve and has repeatedly proven to be a barrier to effective use for scientists unfamiliar with the visual analytics paradigm. Instead, EAM QuickView provides a UI explicitly tailored to the scientific language and reasoning workflows of atmospheric modeling. The UI is intentionally focused on rapid inspection of simulation output. In this sense, EAM QuickView represents the first prototype in a family of ParaView-based analysis UIs that is purpose-built for atmospheric applications. Additional members of this tool family currently under development target workflows such as rapid comparison of multiple simulations and flexible slicing of three-dimensional fields using arbitrary planes and bounding boxes.
 
 # Software design
 
