### 2026-01-29 #9083 @ayenpure
> Apologies [@ayenpure](https://github.com/ayenpure), we missed one section from the list above. In addition to the sections already mentioned, JOSS papers now also require:
>
> * State of the Field (~200 words): A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear "build vs. contribute" justification explaining your unique scholarly contribution and why existing alternatives are insufficient.
>
> So in summary, the requested additions are:
>
> 1. State of the Field (~200 words): A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear "build vs. contribute" justification explaining your unique scholarly contribution and why existing alternatives are insufficient.
> 2. Software Design (~200 words): An explanation of the trade-offs you weighed, the design/architecture you chose, and why it matters for your research application. This should demonstrate meaningful design thinking beyond a superficial code structure description.
> 3. Research Impact Statement (~200 words): Evidence of realized impact (publications, external use, integrations) or credible near-term significance (benchmarks, reproducible materials, community-readiness signals). The evidence should be compelling and specific, not aspirational.
> 4. AI Usage Disclosure: Transparent disclosure of any use of generative AI in the software creation, documentation, or paper authoring. If no AI tools were used, state this explicitly. If AI tools were used, describe how they were used and how the quality and correctness of AI-generated content was verified.
>
> Apologies again for the oversight.

We've made the necessary changes for the paper
