### 2026-04-02 ayenpure/QuickView#47 @kdorheim
Community guidelines
As part of JOSS Review https://github.com/openjournals/joss-reviews/issues/9922#issuecomment-3983383826

While the authors provide guidelines on how community members can report bugs and request features, the community guidelines do not address how community members may or may not contribute to the software.
