### 2026-04-03 ayenpure/QuickView#48 @kdorheim
JOSS Paper Feedback
Related to the JOSS Review: https://github.com/openjournals/joss-reviews/issues/9922#issuecomment-3983383826

The authors do a good job of presenting QuickView and describing its potential usefulness but I do have a few comments and questions. I also think that there are a few elements that should be expanded on before publication.

1. It would be helpful to include citations for E3SM and EAM when they are first mentioned.
2. In L68 the authors mention that EAM QuickView could be adapted for other models by updating the ParaView Reader in the app. Has this been done before?
3. As it is currently written the statement of need clearly makes a case for how useful QuickView is for easily visualizing E3SM/EAM results. Some more details about how it has been adopted and the size of the user community would really strengthen this section.
4. In the state of the field section it would be helpful to mention how QuickView differs from Panoply https://www.giss.nasa.gov/tools/panoply/, another existing interface to plotting ESM netcdf results. Could the authors highlight how QuickView differs from Panoply in terms of capability?
5. Right now the Research Impact Statement is vague and forward looking. Are there any specific manuscripts or model developments the authors could point to that demonstrate how QuickView has been adopted?
