### 2026-04-03 ayenpure/QuickView#49 @kdorheim
Minor Documentation Changes
Related to the documentation section of the JOSS Review: https://github.com/openjournals/joss-reviews/issues/9922#issuecomment-3983383826

For full disclosure, I did not try to install and launch the app following the instructions for app developers. One of the great things about QuickView is how easy it is to install and launch the app and then start to play around with the user interface. One of the nice things was that I did not have to worry about installing dependencies, however it would be nice to have a clear list of the dependencies somewhere for troubleshooting and also to acknowledge the software that QuickView is built upon.
