### 2026-04-03 ayenpure/QuickView#50 @kdorheim
Example Usage
Related to the documentation section of the JOSS Review: https://github.com/openjournals/joss-reviews/issues/9922

While the authors provide files that allow users to play around with the user interface, and QuickView is pretty and fairly intuitive to start using based on how the authors document QuickView’s features, there is no clear example for users to walk through. In the software paper, the authors make the claim that QuickView differs from other netCDF data visualization interfaces in how it lends itself to data analysis. Not only would it be beneficial to users to have a clear example to walk through, but it would also strengthen some of the statements of need and research impact sections of the paper. The most compelling example would demonstrate how to use QuickView for a simple, reproducible analysis.
