### 2026-04-03 #9922 @kdorheim
Hello @ayenpure,

As someone who knows the struggle of plotting netCDF results, I really appreciate how easy and intuitive QuickView is to install and generate quick plots! Although, admittedly, I did not follow the installation and launch instructions for app developers. But I have completed my initial review of the paper and software, which mostly focused on the functionality of the interface and documentation. I have opened several GitHub issues that should be resolved or addressed before I feel comfortable with completing the review checklist.

I do have a question for the editor, @kanishkan91. It does not appear to have any automated tests that check the functionality of the software, but there is a GitHub workflow that confirms that QuickView builds. Could the editor confirm that this is sufficient to check off the automated testing item from the reviewer checklist?

@ayenpure thanks for this contribution, I look forward to learning more about QuickView and seeing it adopted in the future.
